backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.9
djangorestframework==3.16.1
django_celery_beat
gunicorn
httpx
exceptiongroup==1.3.1
kombu==5.6.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ulid==3.1.0
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.3
tzlocal==5.3.1
ulid-py==1.1.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.7.0
